requirements.txt
streamlit==1.31.0
pymupdf==1.23.8
python-docx==1.1.0